requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
django-filter
python-dotenv
django-environ
PyMySQL
gunicorn
reportlab
Pillow
